// src/app/pages/model/memberdetail.ts
// model/memberdetail.ts

export interface Cartail {
    carNumber: string;
    brand: string;
    ownerName: string;
    color: string;
    model: string;
  }
  
  export interface MemberDetail {
    memberNumber: string;
    birthday: string;
    gender: string;
    nickname: string;
    password: string;
    memberLevel: string;
    status: string;
    email: string;
    phone: string;
    maritalStatus: string;
    mobilePhone: string;
    education: string;
    company: string;
    imageFile: string;
    uuid: string;
    remarks: string;
    registrationDate: string;
    registrationUnit: string;
    referrer: string;
    referralCode: string;
  }
  